+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `dmeta/errors.py`
 - `pptx` and `xlsx` support
 - `get_microsoft_format` function in `util.py`
+- `SECURITY.md`
 ### Changed
 - `run_dmeta` in `functions.py`
 - `read_json` in `util.py`