+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - `pptx` and `xlsx` support
 - `get_microsoft_format` function in `util.py`
 - `SECURITY.md`
+- `inplace` flag in clear, clear_all, update and update_all
+- `inplace` flag in CLI
+- `inplace` tests
 ### Changed
 - `run_dmeta` in `functions.py`
 - `read_json` in `util.py`